+++ src/classes/offset_datetime.rs
@@ -406,11 +406,11 @@ fn __sub__(obj_a: PyObj, obj_b: PyObj) -> PyReturn {
 
 #[allow(static_mut_refs)]
 static mut SLOTS: &[PyType_Slot] = &[
-    slotmethod2!(OffsetDateTime, Py_tp_new, __new__),
-    slotmethod2!(OffsetDateTime, Py_tp_str, __str__, 1),
-    slotmethod2!(OffsetDateTime, Py_tp_repr, __repr__, 1),
-    slotmethod2!(OffsetDateTime, Py_tp_richcompare, __richcmp__),
-    slotmethod2!(Py_nb_subtract, __sub__, 2),
+    slotmethod!(OffsetDateTime, Py_tp_new, __new__),
+    slotmethod!(OffsetDateTime, Py_tp_str, __str__, 1),
+    slotmethod!(OffsetDateTime, Py_tp_repr, __repr__, 1),
+    slotmethod!(OffsetDateTime, Py_tp_richcompare, __richcmp__),
+    slotmethod!(Py_nb_subtract, __sub__, 2),
     PyType_Slot {
         slot: Py_tp_doc,
         pfunc: doc::OFFSETDATETIME.as_ptr() as *mut c_void,
@@ -1152,13 +1152,13 @@ static mut METHODS: &[PyMethodDef] = &[
         from_timestamp_nanos,
         doc::OFFSETDATETIME_FROM_TIMESTAMP_NANOS
     ),
-    method_kwargs2!(OffsetDateTime, replace, doc::OFFSETDATETIME_REPLACE),
-    method_kwargs2!(
+    method_kwargs!(OffsetDateTime, replace, doc::OFFSETDATETIME_REPLACE),
+    method_kwargs!(
         OffsetDateTime,
         replace_date,
         doc::OFFSETDATETIME_REPLACE_DATE
     ),
-    method_kwargs2!(
+    method_kwargs!(
         OffsetDateTime,
         replace_time,
         doc::OFFSETDATETIME_REPLACE_TIME
@@ -1168,10 +1168,11 @@ static mut METHODS: &[PyMethodDef] = &[
         parse_strptime,
         doc::OFFSETDATETIME_PARSE_STRPTIME
     ),
-    method_kwargs2!(OffsetDateTime, add, doc::OFFSETDATETIME_ADD),
-    method_kwargs2!(OffsetDateTime, subtract, doc::OFFSETDATETIME_SUBTRACT),
+    method_kwargs!(OffsetDateTime, add, doc::OFFSETDATETIME_ADD),
+    method_kwargs!(OffsetDateTime, subtract, doc::OFFSETDATETIME_SUBTRACT),
     method1!(OffsetDateTime, difference, doc::EXACTTIME_DIFFERENCE),
-    method_kwargs2!(OffsetDateTime, round, doc::OFFSETDATETIME_ROUND),
+    method_kwargs!(OffsetDateTime, round, doc::OFFSETDATETIME_ROUND),
+    classmethod_kwargs!(OffsetDateTime, __get_pydantic_core_schema__, c""),
     PyMethodDef::zeroed(),
 ];